+++ LinqBenchmarks/Array/Int32/ArrayInt32SkipTakeWhere.cs
@@ -73,8 +73,8 @@ namespace LinqBenchmarks.Array.Int32
             var sum = 0;
             foreach (var item in source
                 .ToStructEnumerable()
-                .Skip((uint)Skip)
-                .Take((uint)Count)
+                .Skip(Skip)
+                .Take(Count)
                 .Where(item => item.IsEven()))
                 sum += item;
             return sum;
@@ -87,8 +87,8 @@ namespace LinqBenchmarks.Array.Int32
             var predicate = new Int32IsEven();
             foreach (var item in source
                 .ToStructEnumerable()
-                .Skip((uint)Skip, x=> x)
-                .Take((uint)Count, x=> x)
+                .Skip(Skip, x=> x)
+                .Take(Count, x=> x)
                 .Where(ref predicate, x=>x))
                 sum += item;
             return sum;
