+++ LinqBenchmarks/Enumerable/Int32/EnumerableInt32SkipTakeSelect.cs
@@ -1,5 +1,4 @@
 ﻿using BenchmarkDotNet.Attributes;
-using JM.LinqFaster;
 using NetFabric.Hyperlinq;
 using StructLinq;
 using System.Linq;
@@ -44,8 +43,8 @@ namespace LinqBenchmarks.Enumerable.Int32
             var sum = 0;
             foreach (var item in source
                 .ToStructEnumerable()
-                .Skip((uint)Skip)
-                .Take((uint)Count)
+                .Skip(Skip)
+                .Take(Count)
                 .Select(item => item * 2))
                 sum += item;
             return sum;
@@ -58,8 +57,8 @@ namespace LinqBenchmarks.Enumerable.Int32
             var selector = new DoubleOfInt32();
             foreach (var item in source
                 .ToStructEnumerable()
-                .Skip((uint)Skip, x=> x)
-                .Take((uint)Count, x=>x)
+                .Skip(Skip, x=> x)
+                .Take(Count, x=>x)
                 .Select(ref selector, x=> x, x=>x))
                 sum += item;
             return sum;
