+++ LinqBenchmarks/List/Int32/ListInt32SkipTakeSelect.cs
@@ -2,7 +2,6 @@
 using JM.LinqFaster;
 using NetFabric.Hyperlinq;
 using StructLinq;
-using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -66,8 +65,8 @@ namespace LinqBenchmarks.List.Int32
             var sum = 0;
             foreach (var item in source
                 .ToStructEnumerable()
-                .Skip((uint)Skip)
-                .Take((uint)Count)
+                .Skip(Skip)
+                .Take(Count)
                 .Select(item => item * 2))
                 sum += item;
             return sum;
@@ -80,8 +79,8 @@ namespace LinqBenchmarks.List.Int32
             var selector = new DoubleOfInt32();
             foreach (var item in source
                 .ToStructEnumerable()
-                .Skip((uint)Skip, x=> x)
-                .Take((uint)Count, x=>x)
+                .Skip(Skip, x=> x)
+                .Take(Count, x=>x)
                 .Select(ref selector, x=> x, x=>x))
                 sum += item;
             return sum;
