+++ LinqBenchmarks/List/Int32/ListInt32SkipTakeWhere.cs
@@ -2,7 +2,6 @@
 using JM.LinqFaster;
 using NetFabric.Hyperlinq;
 using StructLinq;
-using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -74,8 +73,8 @@ namespace LinqBenchmarks.List.Int32
             var sum = 0;
             foreach (var item in source
                 .ToStructEnumerable()
-                .Skip((uint)Skip)
-                .Take((uint)Count)
+                .Skip(Skip)
+                .Take(Count)
                 .Where(item => item.IsEven()))
                 sum += item;
             return sum;
@@ -88,8 +87,8 @@ namespace LinqBenchmarks.List.Int32
             var predicate = new Int32IsEven();
             foreach (var item in source
                 .ToStructEnumerable()
-                .Skip((uint)Skip, x=> x)
-                .Take((uint)Count, x=>x)
+                .Skip(Skip, x=> x)
+                .Take(Count, x=>x)
                 .Where(ref predicate, x=>x))
                 sum += item;
             return sum;
