+++ LinqBenchmarks/List/ValueType/ListValueTypeSkipTakeSelect.cs
@@ -2,7 +2,6 @@
 using JM.LinqFaster;
 using NetFabric.Hyperlinq;
 using StructLinq;
-using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -66,7 +65,7 @@ namespace LinqBenchmarks.List.ValueType
             var sum = default(FatValueType);
             foreach (var item in source
                 .ToRefStructEnumerable()
-                .Skip((uint)Skip)
+                .Skip(Skip)
                 .Take(Count)
                 .Select((in FatValueType element) => element * 2))
                 sum += item;
@@ -82,7 +81,7 @@ namespace LinqBenchmarks.List.ValueType
 
             foreach (var item in source
                 .ToRefStructEnumerable()
-                .Skip((uint)Skip, x=>x)
+                .Skip(Skip, x=>x)
                 .Take(Count, x=>x)
                 .Select(ref selector, x=> x, x=>x))
                 sum += item;
